Represent pending/ready members w/ an enum
This is preliminary work to prep for persisting the swarm member's state
to disk to allow resuming from the last known state. Currently when
a storage server starts up, it assumes that it's joining a new swarm/new
members are joining its swarm and does a full message dump to those
members instead of being able to know if a swarm member is new or we
already knew about it and then, choose a synchronisation method that
is more suited for that particular scenario.