Infinite retries with exponential fallback
When a node initiates a recursive swarm request, the initial node awaits
the response from all other nodes before returning to the client.
Children swarm nodes that fail to receive the request are stored into
a retryable request queue to be re-attempted later.

This queue is flushed every 3s by piggybacking onto the swarm member
check function that is periodically invoked by OMQ.